fix: resolve test failures after refactor/lineage removal
- Fix import paths for staged-analysis-engine-refactored in 3 test files
- Update database error message expectations to match actual messages
- Add null safety check for callGraphResult.callEdges in lazy-analysis
